[[ Bug 17737 ]] Screen should be force unlocked after window resize
This patch ensures that the screen is force unlocked after a system
resize window event is processed. This ensures that using 'lock screen'
inside resizeStack does not cause a resize gesture to only be
processed for the first resize.

The reason an explicit force unlock is needed in this instance is that
the system resize events occur in a modal system loop meaning that
the root wait loop is not touched (which does force unlock the screen).

The MCDispatch::configure() method has been removed, and replaced by
MCDispatch::wreshape() - this makes sure the behavior is the same
on all platforms.
