+++ docs/notes/bugfix-17737.md
@@ -1,0 +1,1 @@
+# Screen should be force unlocked after resizeStack message is sent
